make sync test work with older version (credentials were changed) No-Issue
